refactor: reorganize Form & FormPage & modify test
