+++ src/components/Form/Form.test.tsx
@@ -1,9 +1,17 @@
-import React from 'react';
 import { render, screen } from '@testing-library/react';
 import Form from './Form';
+import { FormCardType } from 'components/FormCard/FormCard';
 
 test('renders Form placeholderName', () => {
-  render(<Form />);
+  let cards: FormCardType[] = [];
+
+  function updateCards(newCard: FormCardType) {
+    cards = [...cards, newCard];
+  }
+
+  function showSubmitMessage(): void {}
+
+  render(<Form showSubmitMessage={showSubmitMessage} updateCards={updateCards} />);
   const placeholderName = screen.getByPlaceholderText('Enter your name');
   expect(placeholderName).toBeInTheDocument();
 });
